missed param by dtype in test_slice_replace

+++ pandas/tests/strings/test_strings.py
@@ -476,8 +476,11 @@ def test_slice_mixed_object(start, stop, step, expected):
     ],
 )
 def test_slice_replace(start, stop, repl, expected, any_string_dtype):
-    ser = Series(["short", "a bit longer", "evenlongerthanthat", "", np.nan])
-    expected = Series(expected)
+    ser = Series(
+        ["short", "a bit longer", "evenlongerthanthat", "", np.nan],
+        dtype=any_string_dtype,
+    )
+    expected = Series(expected, dtype=any_string_dtype)
     result = ser.str.slice_replace(start, stop, repl)
     tm.assert_series_equal(result, expected)
